--- 2nd Study/Example/Json/Json.vcxproj.filters
@@ -1,92 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="docs">
-      <UniqueIdentifier>{25b504ff-5a84-4bd1-bde6-16679e21e450}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="docs\Catch">
-      <UniqueIdentifier>{d8faf697-314d-410b-b328-8f9272226e00}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="json">
-      <UniqueIdentifier>{0a9371f7-0ceb-44a7-856a-dff242732a25}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="json\parser">
-      <UniqueIdentifier>{febb6c4b-50cf-4b6b-b1ee-f1ac0f0241fa}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="json\types">
-      <UniqueIdentifier>{78aa2407-f693-415e-91a3-aa6da41f6a6d}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="test">
-      <UniqueIdentifier>{cdbbe9a0-5197-4aa7-ba14-a5cabe3c85ad}</UniqueIdentifier>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="test\test.json_parser.cpp">
-      <Filter>test</Filter>
-    </ClCompile>
-    <ClCompile Include="test\test.json_types.cpp">
-      <Filter>test</Filter>
-    </ClCompile>
-    <ClCompile Include="test\test.main.cpp">
-      <Filter>test</Filter>
-    </ClCompile>
-    <ClCompile Include="json\parser\array_parser.cpp">
-      <Filter>json\parser</Filter>
-    </ClCompile>
-    <ClCompile Include="json\parser\json_builder.cpp">
-      <Filter>json\parser</Filter>
-    </ClCompile>
-    <ClCompile Include="json\parser\object_parser.cpp">
-      <Filter>json\parser</Filter>
-    </ClCompile>
-    <ClCompile Include="json\parser\stream_parser.cpp">
-      <Filter>json\parser</Filter>
-    </ClCompile>
-    <ClCompile Include="json\types\json_array.cpp">
-      <Filter>json\types</Filter>
-    </ClCompile>
-    <ClCompile Include="json\types\json_base.cpp">
-      <Filter>json\types</Filter>
-    </ClCompile>
-    <ClCompile Include="json\types\json_object.cpp">
-      <Filter>json\types</Filter>
-    </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="json\parser.h">
-      <Filter>json</Filter>
-    </ClInclude>
-    <ClInclude Include="json\types.h">
-      <Filter>json</Filter>
-    </ClInclude>
-    <ClInclude Include="json\parser\array_parser.h">
-      <Filter>json\parser</Filter>
-    </ClInclude>
-    <ClInclude Include="json\parser\json_builder.h">
-      <Filter>json\parser</Filter>
-    </ClInclude>
-    <ClInclude Include="json\parser\object_parser.h">
-      <Filter>json\parser</Filter>
-    </ClInclude>
-    <ClInclude Include="json\parser\stream_parser.hpp">
-      <Filter>json\parser</Filter>
-    </ClInclude>
-    <ClInclude Include="json\types\json_array.hpp">
-      <Filter>json\types</Filter>
-    </ClInclude>
-    <ClInclude Include="json\types\json_base.hpp">
-      <Filter>json\types</Filter>
-    </ClInclude>
-    <ClInclude Include="json\types\json_object.hpp">
-      <Filter>json\types</Filter>
-    </ClInclude>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="..\lib\Catch\README.md">
-      <Filter>docs\Catch</Filter>
-    </None>
-    <None Include="..\lib\Catch\docs\tutorial.md">
-      <Filter>docs\Catch</Filter>
-    </None>
-  </ItemGroup>
-</Project>
