--- 2nd Study/Example/Json/json/parser.h
@@ -1,19 +1,0 @@
-// ==== ==== ==== ==== ==== ==== ==== ==== ==== ==== ==== ==== ====
-// 
-//  File    : parser.h
-//  Author  : Park  Dong Ha ( luncliff@gmail.com )
-//  Updated : 16/08/18
-//
-//  Note
-//      Minimal JSON library parser headers 
-//
-// ---- ---- ---- ---- ---- ---- ---- ---- ---- ---- ---- ---- ----
-#ifndef _JSON_PARSER_H_
-#define _JSON_PARSER_H_
-
-#include "./parser/json_builder.h"
-#include "./parser/stream_parser.hpp"
-#include "./parser/array_parser.h"
-#include "./parser/object_parser.h"
-
-#endif
