--- 2nd Study/Example/Json/json/types.h
@@ -1,19 +1,0 @@
-// ==== ==== ==== ==== ==== ==== ==== ==== ==== ==== ==== ==== ====
-// 
-//  File    : types.h
-//  Author  : Park  Dong Ha ( luncliff@gmail.com )
-//  Updated : 16/08/18
-//
-//  Note
-//      Minimal JSON library type headers 
-//
-// ---- ---- ---- ---- ---- ---- ---- ---- ---- ---- ---- ---- ----
-#ifndef _JSON_TYPES_H_
-#define _JSON_TYPES_H_
-
-#include "./types/json_base.hpp"
-#include "./types/json_object.hpp"
-#include "./types/json_array.hpp"
-
-#endif
-
