--- 2nd Study/Example/Json/test/test.main.cpp
@@ -1,14 +1,0 @@
-// ==== ==== ==== ==== ==== ==== ==== ==== ==== ==== ==== ==== ====
-//
-//  File    : test.main.cpp
-//  Updated : 16/08/17
-//  Author  : Park  Dong Ha ( luncliff@gmail.com )
-//
-//  Note
-//      This tells `Catch` library to provide a `main()` here
-//      - only do this in one cpp file
-//
-// ---- ---- ---- ---- ---- ---- ---- ---- ---- ---- ---- ---- ----
-
-#define CATCH_CONFIG_MAIN  
-#include <catch.hpp>
