+++ 2nd Study/Example/README.md
@@ -3,34 +3,28 @@
 > ### 작성자
 >   박 동하 ( luncliff@gmail.com )
 > ### 갱신일
->   16/08/19
+>   16/09/30
 
 
 ## 0. 요약
 2주차 디자인 패턴 스터디에서 사용하는 패턴들이 반영된 예제 코드입니다. 
 
-### `Json`
+### - Json
+
+> 5주차 스터디에서 다시 사용되었습니다.  
+> 해당 폴더를 확인해주세요. :D
+
 Adpater, Bridge, Prototype 패턴을 적용한 코드
 
-### `Nana`
+### - Nana
 Composite 패턴을 적용한 코드
 
 
-## 1. 의존성
+## 1. 외부 라이브러리
 
-### 외부 라이브러리
-
-#### Catch
+### - Catch
 Macro 기반 C++ Testing 라이브러리입니다. 예제코드의 Unit Test를 위해 사용하였습니다.
 
-#### Nana
-[Project Home](), [Repo]()  
+### - Nana
+[Project Home](http://nanapro.org/en-us/), [Repo](https://github.com/cnjinhao/nana)  
 Modern C++를 사용한 GUI 라이브러리입니다.
-
-
-## 2. 코드 설명
-
-### `Json`
-
-
-### `Nana`
