+++ 5th Study/Example/Example.sln
@@ -1,0 +1,34 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Json", "Json\Json.vcxproj", "{2DE18D84-2873-4FFD-BB96-BF9D7592B52B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{434EF148-FBDC-48F7-B16D-096B71E2559A}"
+	ProjectSection(SolutionItems) = preProject
+		.gitignore = .gitignore
+		README.md = README.md
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2DE18D84-2873-4FFD-BB96-BF9D7592B52B}.Debug|x64.ActiveCfg = Debug|x64
+		{2DE18D84-2873-4FFD-BB96-BF9D7592B52B}.Debug|x64.Build.0 = Debug|x64
+		{2DE18D84-2873-4FFD-BB96-BF9D7592B52B}.Debug|x86.ActiveCfg = Debug|Win32
+		{2DE18D84-2873-4FFD-BB96-BF9D7592B52B}.Debug|x86.Build.0 = Debug|Win32
+		{2DE18D84-2873-4FFD-BB96-BF9D7592B52B}.Release|x64.ActiveCfg = Release|x64
+		{2DE18D84-2873-4FFD-BB96-BF9D7592B52B}.Release|x64.Build.0 = Release|x64
+		{2DE18D84-2873-4FFD-BB96-BF9D7592B52B}.Release|x86.ActiveCfg = Release|Win32
+		{2DE18D84-2873-4FFD-BB96-BF9D7592B52B}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
