+++ 5th Study/Example/Json/json.hpp
@@ -1,0 +1,17 @@
+// ==== ==== ==== ==== ==== ==== ==== ==== ==== ==== ==== ==== ====
+// 
+//  File    : json.hpp
+//  Author  : Park  Dong Ha ( luncliff@gmail.com )
+//  Updated : 16/08/18
+//
+//  Note
+//      Minimal JSON library headers 
+//
+// ---- ---- ---- ---- ---- ---- ---- ---- ---- ---- ---- ---- ----
+#ifndef _JSON_HPP_
+#define _JSON_HPP_
+
+#include "./json/types.h"
+#include "./json/parser.h"
+
+#endif
