+++ 5th Study/Example/README.md
@@ -1,0 +1,39 @@
+# 5주차 스터디 예제 코드
+
+> ### 작성자
+>   박 동하 ( luncliff@gmail.com )
+> ### 갱신일
+>   16/09/30 ~ 16/10/01
+
+## 0. 요약
+5주차 디자인 패턴 스터디에서 사용하는 패턴들이 반영된 예제 코드입니다. 
+
+### - Mediator 패턴
+GUI 프로그래밍에서 다수의 Page들간의 관계, 화면전환 코드를 관리하는 것은 꽤나 번거로운 일이 될 수 있습니다.    
+프로그래머의 두뇌를 괴롭히는 이 문제를 `Mediator`로 접근해봅니다.
+
+### - Observer 패턴
+옵저버 패턴으로 간단한 Data-Binding을 구현합니다.   
+옵저버 패턴은 간단하면서도 효율적인 패턴 중 하나인데요, 멀티스레드에서는 어떨까요?
+
+### - State 패턴
+2주차 스터디에서 사용했던 JSON 타입들을 기반으로   
+FSM(Finite State Machine)을 활용한 JSON parser를 구현합니다.
+
+
+## 1. 외부 라이브러리
+
+### - Qt 5.7
+C++ 생태계에서 Qt는 대표적인 Cross Platform 프레임워크로서 자리하고 있습니다. 
+
+예제 코드에서는 GUI를 위해서 사용했지만,   
+Database나 System, Web 프로그래밍에도 유용하게 사용될 수 있는 라이브러리로 가득합니다. 
+
+오픈 소스 개발에는 무료로 사용할 수 있지만 그 외에는 라이센스 비용을 지불해야 합니다.
+
+### - Catch
+Macro 기반 C++ Testing 라이브러리입니다. 예제코드들의 Unit Test를 위해 사용하였습니다.
+
+
+
+
